Moved domTokenToArray to common function
