+++ test/component-helper.js
@@ -43,7 +43,17 @@ const expectComponent = (name, input, expected) => {
   )
 }
 
+function domTokenToArray (obj) {
+  let array = []
+  // iterate backwards ensuring that length is an UInt32
+  for (let i = obj.length >>> 0; i--;) {
+    array[i] = obj[i]
+  }
+  return array
+}
+
 module.exports = {
   expectComponent,
   renderComponentToDom,
+  domTokenToArray,
 }
