+++ test/javascripts/sectors.test.js
@@ -2,6 +2,7 @@
 const jsdom = require('jsdom')
 const { JSDOM } = jsdom
 const Sectors = require('~/src/javascripts/sectors')
+const { domTokenToArray } = require('../component-helper')
 
 const HTML = `
   <html>
@@ -21,15 +22,6 @@ const HTML = `
       </form>
     </body>
   </html>`
-
-function domTokenToArray (obj) {
-  let array = []
-  // iterate backwards ensuring that length is an UInt32
-  for (let i = obj.length >>> 0; i--;) {
-    array[i] = obj[i]
-  }
-  return array
-}
 
 describe('Sector control', function () {
   let document
